+++ src/kvstorehandle.cpp
@@ -17,12 +17,12 @@ void KVStoreHandle::setActiveFileID(std::filesystem::path dataDir) {
         // Should not be if we've called from openStore().
         std::perror("setActiveFileID -> Data directory is empty.");
 
-    uint64_t maxID = 0;
+    uint32_t maxID = 0;
     bool found = false;
 
     for (const auto& datafile : std::filesystem::directory_iterator(dataDir)) {
         if (datafile.path().extension() == ".log") {
-            uint64_t currentID = std::stoul(datafile.path().stem().string());
+            uint32_t currentID = std::stoul(datafile.path().stem().string());
             maxID = std::max(maxID, currentID);
             found = true;
         }
